Started working on some data analysis stuff. Working on improving ripperdoc as while functional he is well
not the best

but thats ok as I am exceited to start working past the "proof of concept page"

The next step is the "soulkiller" library. Basically a method to analyze training data and something that will be added to ripper doc to allow the user to select training data and whatnot.